+++ Libplanet.Benchmarks/SwarmBenchmark.cs
@@ -58,7 +58,7 @@ namespace Libplanet.Benchmarks
             _blockChains = new BlockChain<DumbAction>[SwarmNumber];
             _swarms = new Swarm<DumbAction>[SwarmNumber];
 
-            var genesisBlock = BlockChain<DumbAction>.MakeGenesisBlock();
+            var genesisBlock = BlockChain<DumbAction>.MakeGenesisBlock(_hashAlgorithm);
             var tasks = new List<Task>();
             for (int i = 0; i < SwarmNumber; i++)
             {